Create Table_Printer
At first, it will tell you to input the number. the number of that you want to print table. After taking input it will print the table of that number. And at last, it will ask you "Do you want to continue?" if yes then you can repeat, and if you input "No" then the program will break there.